backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
# psycopg2-binary==2.9.9  # Optional - Django will use SQLite if not available
python-dotenv==1.0.0
Pillow>=10.3.0
djangorestframework-simplejwt==5.3.1
# AI Grading (optional - install one or both for Writing grading)
openai>=1.0.0  # For OpenAI GPT-4 grading
anthropic>=0.18.0  # For Anthropic Claude grading

